first stab.
will need to re-key on the entity.name not the entity.email so that the
deduping works correctly -- what if an exploder template has the same
entity in the explode partytype as well as a regular to partytype? the
dedup needs to happen at the entity.name level before we proceed to
label with entity._es_num.